% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Compute Cosine Similarity Between Two Vectors}
\usage{
cosine_similarity(vec1, vec2)
}
\arguments{
\item{vec1}{Numeric vector}

\item{vec2}{Numeric vector}
}
\value{
Cosine similarity between two vectors
}
\description{
Calculates the cosine similarity between two numeric vectors, which measures the
angle between them in the vector space (i.e. semantic relationship/similarity between text)
}
