% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{get_text_embedding}
\alias{get_text_embedding}
\title{Generate Text Embedding via Ollama}
\usage{
get_text_embedding(text, model = "mxbai-embed-large")
}
\arguments{
\item{text}{A character string of the text to embed.}

\item{model}{The model to use for generating the embedding. Default is "mxbai-embed-large".}
}
\value{
A numeric vector representing the text embedding.
}
\description{
Gets vector representation (embedding) for the input text using the specified model.
}
