% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{get_top_matches}
\alias{get_top_matches}
\title{Return Top Matching Vector and its Index Based on Prompt Vector}
\usage{
get_top_matches(prompt_vec, top_n = 3, embedding_matrix_df, corpus_text_df)
}
\arguments{
\item{prompt_vec}{Numeric vector of prompt supplied by return_top_matches function}

\item{top_n}{Number of top matches to return.}

\item{embedding_matrix_df}{A data frame of embedding vectors (unnested)}

\item{corpus_text_df}{A data frame of text content and metadata (e.g., libguide text).}
}
\value{
A data frame containing necessary data for top results.
}
\description{
Given a prompt, returns the top \code{n} matching vectors from a provided vector space
(unnested vector store) based on cosine similarity of embedding vectors.
}
