% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{return_top_matches}
\alias{return_top_matches}
\title{Return Top Matching Text Entries Based on Prompt}
\usage{
return_top_matches(prompt, top_n, embedding_matrix_df, corpus_text_df)
}
\arguments{
\item{prompt}{A user-supplied character string.}

\item{top_n}{Number of top matches to return.}

\item{embedding_matrix_df}{A data frame of embedding vectors (unnested)}

\item{corpus_text_df}{A data frame of text content and metadata (e.g., libguide text).}
}
\value{
Printed formatted string output of the top matches.
}
\description{
Given a prompt, returns the top \code{n} matching entries from a provided text dataset
based on cosine similarity of embedding vectors; includes text wrapper containing
appropriate titles and URLs.
}
